# Fix issues with caching static files.