--- a/src/controllers/root_controller.js
+++ b/src/controllers/root_controller.js
@@ -23,7 +23,7 @@ RootController.get('/', TunnelSetup.isTunnelSet,
   function(request, response) {
     response.sendFile('index.html', {
       root: Constants.VIEWS_PATH,
-      maxAge: '14d'
+      maxAge: 0,
     });
   }
 );