@hobinjk 2018-01-18
The contributing factor to the 15 second loading time isn't throughput, it's latency. The ETag header prevents the file from being sent but still costs a full round trip. Using maxAge prevents the need for the round trip but means that you have to use a hard refresh to signal that you don't want to use the cache.

In terms of real-world performance:
[WebPageTest of master, 15s -> 3.7s repeat](https://www.webpagetest.org/result/180118_HF_095a6a536221460f46a9a1e7a08979d8/)
[WebPageTest of this PR, 15s -> 14.1s repeat](https://www.webpagetest.org/result/180118_5C_3fc1c8794b2f4e66b678f2703455f7ac/)

In general, I have yet to have any issues with overzealous caching so I don't really see where #454 is coming from. I'll put a comment describing my workflow in that issue in case it helps.