@mrstegeman 2018-01-18
Yikes! @dhylands and I have both seen this issue several times, though, especially when modifying some of the static JS files. It probably won't be a huge problem for our users (other than after upgrades, for up to 2 weeks), but it certainly makes development a pain.