@hobinjk 2018-01-18
A ServiceWorker would be able to properly fix this, serving from the cache initially and updating in the background, e.g. https://serviceworke.rs/strategy-cache-and-update.html

#447 exists to track writing a SW generator and is currently up for grabs if you'd like to work on that.