@dhylands 2018-01-18
Serving JS from the cache when its mismatched can cause crashes. I've seen this once on my RPi where I had a previous version installed, and then updated the software, but had some of the JS coming from the cache and making REST requests which no longer existed rather than using the new code which would have used the new REST requests.